lang: Cleanup language install scripts

+++ data/lang_patch.sh
@@ -3,41 +3,43 @@
 [ -e "/tmp/lang_patch.log" ] && return 0
 
 DIR_PATCH=/etc/crontabs/patches
+DIR_BACKUP=$DIR_PATCH/lang_backup
 
-if [ `ls $DIR_PATCH/base.*.lmo |wc -l` -eq 0 ]; then
+if [ `find $DIR_PATCH -maxdepth 1 -name 'base.*.lmo' | wc -l` -eq 0 ]; then
 	return 0
 fi
 
-#if [ -e "/etc/xiaoqiang_version" ]; then
-#	return 0
-#fi
-
-if [ -e "/usr/lib/lua/luci/i18n/base.en.lmo" ]; then
-	return 0
+if ! mount | grep -q ' on /usr/lib/lua/luci' ; then
+	mkdir -p /tmp/_usr_lib_lua_luci
+	cp -rf /usr/lib/lua/luci/* /tmp/_usr_lib_lua_luci/
+	mount --bind /tmp/_usr_lib_lua_luci /usr/lib/lua/luci
+fi
+if ! mount | grep -q ' on /usr/lib/lua/luci' ; then
+	return 1  # error
+fi
+if [ ! -f /tmp/_usr_lib_lua_luci/i18n.lua ]; then
+	return 1  # error
 fi
 
-mkdir -p /tmp/_usr_lib_lua_luci
-cp -rf /usr/lib/lua/luci/* /tmp/_usr_lib_lua_luci/
-mount --bind /tmp/_usr_lib_lua_luci /usr/lib/lua/luci
+if ! mount | grep -q ' on /usr/share/xiaoqiang' ; then
+	mkdir -p /tmp/_usr_share_xiaoqiang
+	cp -rf /usr/share/xiaoqiang/* /tmp/_usr_share_xiaoqiang/
+	mount --bind /tmp/_usr_share_xiaoqiang /usr/share/xiaoqiang
+fi
+if ! mount | grep -q ' on /usr/share/xiaoqiang' ; then
+	return 1  # error
+fi
+if [ ! -f /tmp/_usr_share_xiaoqiang/xiaoqiang_version ]; then
+	return 1  # error
+fi
 
-cp $DIR_PATCH/base.*.lmo /usr/lib/lua/luci/i18n
-
-# save original file
-cp -f /usr/share/xiaoqiang/xiaoqiang_version /etc/xiaoqiang_version
-
-mkdir -p /tmp/_usr_share_xiaoqiang
-cp -rf /usr/share/xiaoqiang/* /tmp/_usr_share_xiaoqiang/
-mount --bind /tmp/_usr_share_xiaoqiang /usr/share/xiaoqiang
+cp -f $DIR_PATCH/base.*.lmo /usr/lib/lua/luci/i18n/
 
 # unlock WEB lang menu
 sed -i 's/ and features\["system"\]\["i18n"\] == "1" //' /usr/lib/lua/luci/view/web/inc/sysinfo.htm
 
 # unlock change luci.main.lang
 sed -i "s/option CHANNEL 'stable'/option CHANNEL 'release'/g" /usr/share/xiaoqiang/xiaoqiang_version
-
-if [ -e "$DIR_PATCH/lang_patch1.sh" ]; then
-	sh $DIR_PATCH/lang_patch1.sh
-fi
 
 echo "lang patched" > /tmp/lang_patch.log
 
@@ -48,5 +50,7 @@ uci set luci.languages.en=English
 uci commit luci
 
 # reload luci
-luci-reload & rm -f /tmp/luci-indexcache & luci-reload
+luci-reload
+rm -f /tmp/luci-indexcache
+luci-reload
